+++ src/entries/mod.rs
@@ -1,5 +1,6 @@
 mod entry;
 mod filter;
+pub mod input;
 mod sequence;
 
 use crate::utils;
